feat: handle empty

+++ pidfile.go
@@ -9,6 +9,8 @@ import (
 )
 
 var (
+	RaiseEmptyPath bool
+
 	ErrNoPIDFile  = errors.New("no pidfile")
 	ErrDuplicated = errors.New("duplicated")
 )
@@ -37,6 +39,12 @@ func (pf *PIDFile) Generate() interface{ Cleanup() error } {
 	if pf == nil {
 		panic(ErrNoPIDFile)
 	}
+	if pf.path == "" {
+		if RaiseEmptyPath {
+			panic(ErrNoPIDFile)
+		}
+		return pf
+	}
 	if atomic.CompareAndSwapInt32(&pf.generate, 0, 1) {
 		for ix := range pf.cons {
 			if err := pf.cons[ix](); err != nil {
@@ -63,6 +71,12 @@ func (pf *PIDFile) Generate() interface{ Cleanup() error } {
 func (pf *PIDFile) Cleanup() error {
 	if pf == nil {
 		return ErrNoPIDFile
+	}
+	if pf.path == "" {
+		if RaiseEmptyPath {
+			return ErrNoPIDFile
+		}
+		return nil
 	}
 	if atomic.LoadInt32(&pf.generate) == 1 && atomic.CompareAndSwapInt32(&pf.cleanup, 0, 1) {
 		return os.Remove(pf.path)
